Move update-all-blocks.js to libraries
