+++ addons/custom-block-shape/userscript.js
@@ -1,4 +1,4 @@
-import { updateAllBlocks } from "./update-all-blocks.js";
+import { updateAllBlocks } from "../../libraries/common/cs/update-all-blocks.js";
 
 export default async function ({ addon, console }) {
   var BlocklyInstance = await addon.tab.traps.getBlockly();
