+++ addons/editor-square-inputs/userscript.js
@@ -1,4 +1,4 @@
-import { updateAllBlocks } from "../custom-block-shape/update-all-blocks.js";
+import { updateAllBlocks } from "../../libraries/common/cs/update-all-blocks.js";
 
 export default async function ({ addon }) {
   const ScratchBlocks = await addon.tab.traps.getBlockly();
